Try to get the server to work
It doesn't seem interested. It won't work unless you put in index.php
first in the URL, but I don't want to do that. I know it can be done,
but I can't figure out how. It gives me this error in the log whenever I
get a page after it should not make me put in index.php:
Request exceeded the limit of 10 internal redirects due to probable
configuration error. Use 'LimitInternalRecursion' to increase the limit
if necessary. Use 'LogLevel debug' to get a backtrace